AMPOFFLINE-506 Using bunyan.child and prefixing with module name

+++ app/modules/util/LoggerManager.js
@@ -10,47 +10,51 @@ import {
 import LoggerSettings from '../../utils/LoggerSettings';
 import FileManager from './FileManager';
 
-
 /* To understand the levels: https://github.com/trentm/node-bunyan#levels
  * 30: info
  * 40: warn
  * 50: error */
 export default class LoggerManager {
+  static bunyanLog = null;
 
-  static bunyanLog = LoggerManager.initialize();
+  static getBunyanLog() {
+    if (!this.bunyanLog) {
+      console.log('initialize');
+      fs.readdirSync(LOG_DIR).sort().reverse().slice(NR_LOG_FILES)
+        .forEach(filename =>
+          fs.unlink(path.join(LOG_DIR, filename)));
 
-  static initialize() {
-    console.log('initialize');
-    fs.readdirSync(LOG_DIR).sort().reverse().slice(NR_LOG_FILES)
-      .forEach(filename =>
-               fs.unlink(path.join(LOG_DIR, filename)));
-
-    const settings = LoggerSettings.getDefaultConfig(process.env.NODE_ENV);
-    FileManager.createDataDir(LOG_DIR);
-    const date = new Date();
-    let file = `${LOG_FILE_NAME}.${date.toJSON().replace(/:|\./g, '-')}.${LOG_FILE_EXTENSION}`;
-    file = FileManager.getFullPath(LOG_DIR, file);
-    const log = bunyan.createLogger({
-      name: 'amp',
-      streams: [{ level: settings.level, path: file }
-      ]
-    });
-    return log;
+      const settings = LoggerSettings.getDefaultConfig(process.env.NODE_ENV);
+      FileManager.createDataDir(LOG_DIR);
+      const date = new Date();
+      let file = `${LOG_FILE_NAME}.${date.toJSON().replace(/:|\./g, '-')}.${LOG_FILE_EXTENSION}`;
+      file = FileManager.getFullPath(LOG_DIR, file);
+      this.bunyanLog = bunyan.createLogger({
+        name: 'amp',
+        streams: [{ level: settings.level, path: file }
+        ]
+      });
+    }
+    return this.bunyanLog;
   }
 
-  static log(message) {
-    this.bunyanLog.info(message);
+  constructor(module) {
+    this.logger = this.constructor.getBunyanLog().child({ module });
   }
 
-  static debug(message) {
-    this.bunyanLog.debug(message);
+  log(message) {
+    this.logger.info(message);
   }
 
-  static warn(message) {
-    this.bunyanLog.warn(message);
+  debug(message) {
+    this.logger.debug(message);
   }
 
-  static error(message) {
-    this.bunyanLog.error(message);
+  warn(message) {
+    this.logger.warn(message);
+  }
+
+  error(message) {
+    this.logger.error(message);
   }
 }
